# Avoid an Arc::clone per row in benchmark